requirements.txt
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
ConfigArgParse==1.5.3
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==0.9.1
fastapi==0.75.0
fastjsonschema==2.16.1
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==2.0
gevent==21.12.0
geventhttpclient==1.5.5
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
locust==2.10.1
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
ninja==1.10.2.3
notebook==6.4.12
numpy==1.23.1
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.2.7
onnxruntime==1.8.0
opencv-python-headless==4.2.0.32
orjson==3.7.8
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.28.1
rich==12.5.1
roundrobin==0.0.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
starlette==0.17.1
streamlit==1.11.0
tabulate==0.8.10
terminado==0.15.0
thop==0.1.1.post2207130030
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
torch==1.10.2+cpu
torchvision==0.11.3+cpu
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==4.3.0
urllib3==1.26.10
uvicorn==0.15.0
uvloop==0.16.0
validators==0.20.0
watchdog==2.1.9
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
